Add files via upload
This notebook contains the code used for analyzing kidney stone images with the ultimate goal of classifying six different types of stones. We applied a transfer learning approach using the ResNet 50 architecture and ImageNet weights on a clinical dataset. Initially, the model is trained to distinguish between kidney stones in dataset 1. Subsequently, it will be retrained to categorize renal calculi from dataset 2.